feat: More UI responsiveness
- Collapse Run list into drawer on narrow screens
- Move Runs, Machines and Services to horizontal tabs
- Standardise border radius'
