+++ app/app/clusters/[clusterId]/runs/page.tsx
@@ -40,7 +40,7 @@ export default async function Page({
 
   return (
     <div className="flex flex-col overflow-auto px-2 space-y-4">
-      <Card className="w-full onboarding-prompt bg-white border border-gray-200 rounded-xl transition-all duration-200 hover:shadow-md mb-6">
+      <Card className="w-full onboarding-prompt bg-white border border-gray-200 rounded-xl transition-all duration-200 mb-6">
         <CardContent className="pt-6">
           <PromptTextarea clusterId={clusterId} />
         </CardContent>
